+++ src/templates/vote-page.js
@@ -3,6 +3,11 @@ import { graphql } from 'gatsby'
 import { DragDropContext } from 'react-beautiful-dnd'
 import styled from 'styled-components'
 import List from '../components/list'
+import {
+    initialCandidateData,
+    initialPartyData,
+    initialOrderData,
+} from '../data/initial-data'
 import Result from '../components/result'
 import Layout from '../components/layout'
 
@@ -37,77 +42,13 @@ const ColResult = styled.div`
     }
 `
 
-// Cells / candidate preferences
-const initialData = {
-    'dingo-party-bob-doggy': {
-        given: 'Bob',
-        surname: 'Doggy',
-        id: 'dingo-party-bob-doggy',
-        partyId: 'dpa',
-        description: 'Some desc here blah',
-        order: 1,
-    },
-    'dingo-party-frank-kangaroo': {
-        given: 'Frank',
-        surname: 'Kangaroo',
-        id: 'dingo-party-frank-kangaroo',
-        partyId: 'dpa',
-        description: 'Some desc here blah',
-        order: 2,
-    },
-    'akp-sue-koala': {
-        given: 'Sue',
-        surname: 'Koala',
-        id: 'dingo-party-sue-koala',
-        partyId: 'akp',
-        description: 'Some desc here blah',
-        order: 3,
-    },
-    'akp-marg-koala': {
-        given: 'Marg',
-        surname: 'Koala',
-        id: 'dingo-party-marg-koala',
-        partyId: 'akp',
-        description: 'Some desc here blah',
-        order: 4,
-    },
-    'akp-francois-budgie': {
-        given: 'Francois',
-        surname: 'Budgie',
-        id: 'dingo-party-francois-budgie',
-        partyId: 'akp',
-        description: 'Some desc here blah',
-        order: 5,
-    },
-}
-
-// Columns
-const initialPartyData = {
-    akp: {
-        name: 'Australian Koala Party',
-        candidates: ['akp-francois-budgie', 'akp-marg-koala', 'akp-sue-koala'],
-    },
-    dpa: {
-        name: 'The dingo party of Australia',
-        candidates: ['dingo-party-bob-doggy', 'dingo-party-frank-kangaroo'],
-    },
-}
-
-const initialOrder = [
-    'dingo-party-bob-doggy',
-    'dingo-party-frank-kangaroo',
-    'akp-sue-koala',
-    'akp-francois-budgie',
-    'akp-marg-koala',
-]
-
 export default class App extends Component {
     constructor() {
         super()
         this.state = {
-            data: initialData,
+            data: initialCandidateData,
             partyData: initialPartyData,
-            order: initialOrder,
+            order: initialOrderData,
         }
         this.onDragEnd = this.onDragEnd.bind(this)
     }
@@ -166,6 +107,7 @@ export default class App extends Component {
     }
 }
 
+// To Do - Wire in live data
 // export const query = graphql`
 //     query($slug: String!) {
 //         statesYaml(fields: { slug: { eq: $slug } }) {
